+++ src/main/java/tqs/sparkflow/stationService/exception/ErrorResponse.java
@@ -1,0 +1,39 @@
+package tqs.sparkflow.stationService.exception;
+
+import java.time.LocalDateTime;
+
+public class ErrorResponse {
+    private int status;
+    private String error;
+    private String message;
+    private String path;
+    private LocalDateTime timestamp;
+
+    public ErrorResponse(int status, String error, String message, String path) {
+        this.status = status;
+        this.error = error;
+        this.message = message;
+        this.path = path;
+        this.timestamp = LocalDateTime.now();
+    }
+
+    public int getStatus() {
+        return status;
+    }
+
+    public String getError() {
+        return error;
+    }
+
+    public String getMessage() {
+        return message;
+    }
+
+    public String getPath() {
+        return path;
+    }
+
+    public LocalDateTime getTimestamp() {
+        return timestamp;
+    }
+}
