--- src/main/java/tqs/sparkflow/station_service/StationServiceApplication.java
@@ -1,16 +1,0 @@
-package tqs.sparkflow.station_service;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-/**
- * Main application class for the Station Service.
- * This is the entry point for the Spring Boot application.
- */
-@SpringBootApplication
-public class StationServiceApplication {
-
-  public static void main(String[] args) {
-		SpringApplication.run(StationServiceApplication.class, args);
-	}
-}
