+++ src/test/java/tqs/sparkflow/stationService/TestStationServiceApplication.java
@@ -1,0 +1,13 @@
+package tqs.sparkflow.stationService;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+import tqs.sparkflow.stationService.StationServiceApplication;
+
+@SpringBootTest(classes = StationServiceApplication.class)
+class TestStationServiceApplication {
+
+	@Test
+	void contextLoads() {
+	}
+}
