+++ src/test/java/tqs/sparkflow/stationService/service/OpenChargeMapServiceInvalidKeyIT.java
@@ -1,0 +1,50 @@
+package tqs.sparkflow.stationService.service;
+
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.ActiveProfiles;
+import org.springframework.test.context.TestPropertySource;
+import tqs.sparkflow.stationService.StationServiceApplication;
+import tqs.sparkflow.stationService.config.TestConfig;
+import tqs.sparkflow.stationService.repository.StationRepository;
+
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
+
+@SpringBootTest(
+    classes = {StationServiceApplication.class, TestConfig.class},
+    properties = {
+        "spring.main.allow-bean-definition-overriding=true"
+    }
+)
+@ActiveProfiles("test")
+@TestPropertySource(properties = {
+    "openchargemap.api.key=invalid-test-key"
+})
+class OpenChargeMapServiceInvalidKeyIT {
+
+    @Autowired
+    private OpenChargeMapService openChargeMapService;
+
+    @Autowired
+    private StationRepository stationRepository;
+
+    @BeforeEach
+    void setUp() {
+        stationRepository.deleteAll();
+    }
+
+    @Test
+    void whenPopulatingStationsWithInvalidApiKey_thenThrowsException() {
+        // Given
+        double latitude = 38.7223;
+        double longitude = -9.1393;
+        int radius = 10;
+
+        // When/Then
+        assertThatThrownBy(() -> openChargeMapService.populateStations(latitude, longitude, radius))
+            .isInstanceOf(IllegalStateException.class)
+            .hasMessageContaining("Access denied to Open Charge Map API");
+    }
+}
