--- src/test/java/tqs/sparkflow/station_service/cucumber/CucumberTestConfig.java
@@ -1,38 +1,0 @@
-package tqs.sparkflow.station_service.cucumber;
-
-import org.springframework.boot.test.web.client.TestRestTemplate;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Profile;
-import org.springframework.core.env.Environment;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.web.SecurityFilterChain;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
-
-@Configuration
-@EnableWebSecurity
-@Profile("test")
-@ConditionalOnProperty(name = "spring.profiles.active", havingValue = "test")
-public class CucumberTestConfig {
-    
-    @Bean
-    public String localServerPort(Environment environment) {
-        return environment.getProperty("local.server.port", "0");
-    }
-
-    @Bean
-    public TestRestTemplate testRestTemplate() {
-        return new TestRestTemplate();
-    }
-
-    @Bean
-    public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
-        http
-            .csrf(csrf -> csrf.disable())
-            .authorizeHttpRequests(auth -> auth
-                .anyRequest().permitAll()
-            );
-        return http.build();
-    }
-}
