--- src/test/java/tqs/sparkflow/station_service/cucumber/StationStepDefinitions.java
@@ -1,92 +1,0 @@
-package tqs.sparkflow.station_service.cucumber;
-
-import io.cucumber.java.en.Given;
-import io.cucumber.java.en.Then;
-import io.cucumber.java.en.When;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.web.client.TestRestTemplate;
-import org.springframework.boot.test.web.server.LocalServerPort;
-import org.springframework.core.ParameterizedTypeReference;
-import org.springframework.http.HttpMethod;
-import org.springframework.http.ResponseEntity;
-import tqs.sparkflow.station_service.model.Station;
-import tqs.sparkflow.station_service.repository.StationRepository;
-
-import java.util.List;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-public class StationStepDefinitions {
-
-    @LocalServerPort
-    private int port;
-
-    @Autowired
-    private TestRestTemplate restTemplate;
-
-    @Autowired
-    private StationRepository stationRepository;
-
-    private ResponseEntity<List<Station>> response;
-    private ResponseEntity<Station> singleResponse;
-
-    @Given("there are stations in the system")
-    public void thereAreStationsInTheSystem() {
-        // Add test stations to the repository
-        Station station1 = new Station();
-        station1.setId("1");
-        station1.setName("Test Station 1");
-        stationRepository.save(station1);
-
-        Station station2 = new Station();
-        station2.setId("2");
-        station2.setName("Test Station 2");
-        stationRepository.save(station2);
-    }
-
-    @When("I request all stations")
-    public void iRequestAllStations() {
-        response = restTemplate.exchange(
-            "http://localhost:" + port + "/stations",
-            HttpMethod.GET,
-            null,
-            new ParameterizedTypeReference<List<Station>>() {}
-        );
-    }
-
-    @Then("I should receive a list of stations")
-    public void iShouldReceiveAListOfStations() {
-        assertNotNull(response, "Response should not be null");
-        assertEquals(200, response.getStatusCode().value(), "Status code should be 200");
-        List<Station> body = response.getBody();
-        assertNotNull(body, "Response body should not be null");
-        assertFalse(body.isEmpty(), "Response body should not be empty");
-        assertEquals(2, body.size(), "Should have 2 stations in the response");
-    }
-
-    @Given("there is a station with ID {string}")
-    public void thereIsAStationWithId(String id) {
-        Station station = new Station();
-        station.setId(id);
-        station.setName("Test Station " + id);
-        stationRepository.save(station);
-    }
-
-    @When("I request the station with ID {string}")
-    public void iRequestTheStationWithId(String id) {
-        singleResponse = restTemplate.getForEntity(
-            "http://localhost:" + port + "/stations/" + id,
-            Station.class
-        );
-    }
-
-    @Then("I should receive the station details")
-    public void iShouldReceiveTheStationDetails() {
-        assertNotNull(singleResponse, "Response should not be null");
-        assertEquals(200, singleResponse.getStatusCode().value(), "Status code should be 200");
-        Station body = singleResponse.getBody();
-        assertNotNull(body, "Response body should not be null");
-        assertEquals("1", body.getId(), "Station ID should match");
-        assertEquals("Test Station 1", body.getName(), "Station name should match");
-    }
-}
